+++ dx/src/exec/libdx/cache.c
@@ -35,7 +35,7 @@ static struct cache {
 } *cache;
 
 /* prototypes */
-static int cachescavenger(unsigned int n);
+static int cachescavenger(ulong n);
 static struct entry *find(char *fun, int key, int n, Object *in);
 static void remove_(struct entry *this); /* '_' resolves os2 name conflict */
 static void insert(struct entry *this);
@@ -51,7 +51,7 @@ Object DXGetCacheEntry(char *fun, int key, int n, ...);
 /* cache reclaim routine
  */
 static int
-cachescavenger(unsigned int n)
+cachescavenger(ulong n)
 {
     struct entry *this;
